refactor: 人口に関する型定義ファイルが増えたため、ファイル名とパスを変更
