+++ src/components/Graph/PopulationGraph.tsx
@@ -1,11 +1,12 @@
 'use client';
 
-import type { PopulationDataPoint, PrefectureData } from '@/types/population';
+import type { PopulationDataPoint, PrefectureData } from '@/types/population/populationData';
 import { PopulationTypeLabels } from '@/types/population/populationType';
 import { fetchPerYearPopulation } from '@/utils/population/fetchPerYearPopulation';
 import Highcharts from 'highcharts';
 import HighchartsReact from 'highcharts-react-official';
 import { useEffect, useRef, useState } from 'react';
+
 type PopulationGraphProps = {
     prefectureData: PrefectureData[]; // 選択された都道府県データ、prefCode と 都道府県名の組の配列
     dataType: number; // 表示する人口データの種類 (0: 総人口, 1: 年少人口, 2: 生産年齢人口, 3: 老年人口)
